src/cloud-functions/causal_salida_pred/requirements.txt
google-api-python-client
google-cloud-storage
google-cloud-bigquery
nltk
tensorflow